Create Map (#10)
* Create new map using hello world text. Also confirmation dialog for IA delete

* PR updates, also missed a new file

* Take in map name second to avoid race conditions on validation

+++ src/commands/integrationAccountCommands.ts
@@ -3,22 +3,31 @@
  *  Licensed under the MIT License. See LICENSE in the project root for license information.
  *--------------------------------------------------------------------------------------------*/
 
+import * as vscode from "vscode";
 import { AzureTreeDataProvider, IAzureNode } from "vscode-azureextensionui";
 import { localize } from "../localize";
 import { IntegrationAccountTreeItem } from "../tree/integration-account/IntegrationAccountTreeItem";
 import { openAndShowTextDocument } from "../utils/commandUtils";
+import { DialogResponses } from "../utils/dialogResponses";
 
 export async function deleteIntegrationAccount(tree: AzureTreeDataProvider, node?: IAzureNode): Promise<void> {
     if (!node) {
         node = await tree.showNodePicker(IntegrationAccountTreeItem.contextValue);
     }
 
-    await node.runWithTemporaryDescription(
-        localize("azIntegrationAccounts.deleting", "Deleting..."),
-        async () => {
-            await node!.deleteNode();
-        }
-    );
+    const result = await vscode.window.showWarningMessage(
+        localize("azIntegrationAccounts.deleteIntegrationAccountPrompt", "Are you sure that you want to delete the whole integration account? This can not be undone."),
+        DialogResponses.yes,
+        DialogResponses.no);
+
+    if (result === DialogResponses.yes) {
+        await node.runWithTemporaryDescription(
+            localize("azIntegrationAccounts.deleting", "Deleting..."),
+            async () => {
+                await node!.deleteNode();
+            }
+        );
+    }
 }
 
 export async function viewIntegrationAccountProperties(tree: AzureTreeDataProvider, node?: IAzureNode): Promise<void> {
